+++ web/gds-user-ui/src/application/routes/routes.ts
@@ -14,7 +14,7 @@ import MembershipGuide from 'components/Section/MembershipGuide';
 import IntegrateAndComply from 'components/Section/IntegrateAndComply';
 import CertificateManagement from '../../components/CertificateManagement';
 import VerifyPage from 'modules/verify';
-import Collaborators from 'modules/dashboard/collaborators/Collaborators';
+import Collaborators from 'modules/dashboard/collaborator';
 
 const Overview = lazy(() => import('modules/dashboard/overview'));
 const CertificateRegistrationPage = lazy(
