+++ web/gds-user-ui/src/application/store/selectors/user.ts
@@ -1,0 +1,6 @@
+import { RootStateOrAny } from 'react-redux';
+import { createSelector } from 'reselect';
+
+const rootState = (state: RootStateOrAny) => state.user;
+
+export const getCurrentStep = createSelector(rootState, (state) => state.currentStep);
